+++ src/TheNextWeek/hittable_list.h
@@ -52,7 +52,7 @@ class hittable_list : public hittable {
 
         for (const auto& object : objects) {
             if (!object->bounding_box(time_start, time_end, temp_box)) return false;
-            output_box = first_box ? temp_box : surrounding_box(output_box, temp_box);
+            output_box = first_box ? temp_box : aabb(output_box, temp_box);
             first_box = false;
         }