+++ src/TheNextWeek/moving_sphere.h
@@ -59,7 +59,7 @@ class moving_sphere : public hittable {
         aabb box1(
             center(time_end) - vec3(radius, radius, radius),
             center(time_end) + vec3(radius, radius, radius));
-        output_box = surrounding_box(box0, box1);
+        output_box = aabb(box0, box1);
         return true;
     }