+++ src/common/interval.h
@@ -11,10 +11,12 @@
 
 class interval {
   public:
-    double min, max;
+    interval() : min(+infinity), max(-infinity) {} // Default interval is empty
 
     interval(double _min, double _max) : min(_min), max(_max) {}
-    interval() : min(+infinity), max(-infinity) {} // Default interval is empty
+
+    interval(const interval& a, const interval& b)
+      : min(fmin(a.min, b.min)), max(fmax(a.max, b.max)) {}
 
     bool contains(double x) const {
         return min <= x && x <= max;
@@ -25,6 +27,15 @@ class interval {
         if (x > max) return max;
         return x;
     }
+
+    interval& operator+=(double offset) {
+        min += offset;
+        max += offset;
+        return *this;
+    }
+
+  public:
+    double min, max;
 };
 
 const static interval empty   (+infinity, -infinity);